CDFNet/requirements.txt
addict
albumentations==1.1.0
catalyst==20.9
einops
matplotlib
numpy
opencv-python
packaging
pytorch-wavelets==1.3.0
pytorch-lightning==1.5.9
scipy
timm
tqdm
ttach









